refactor _dicom, _icom, and _mosaiq out
